+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `Breadcrumbs::update` to update an existing crumb
 * Added the events: `breadcrumbs:beforeUpdate` and `breadcrumbs:afterUpdate`
 * Updated `Breadcrumbs::log` in order to add the ability to catch the exception in your custom listener
+* Detect empty `Breadcrumbs::$elements` on update or remove
 
 # [1.0.0](https://github.com/phalcongelist/breadcrumbs/releases/tag/v1.0.0) (2016-02-21)
 * Initial release